% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addin.R
\name{rs_fn_defn_jump}
\alias{rs_fn_defn_jump}
\title{Jump to a function definition in the local project}
\usage{
rs_fn_defn_jump(context = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{context}{the active document context from {rstudioapi}}
}
\value{
nothing
}
\description{
Search for a local function definion matching the function the cursor is
on, and jump to it, if found.
}
\author{
Miles McBain
}
