% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addin.R
\name{rs_fnmate}
\alias{rs_fnmate}
\title{rs_fnmate}
\usage{
rs_fnmate(context = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{context}{from the rstudioapi}
}
\value{
nothing.
}
\description{
Generate a function definition file from a function call in RStudio
}
