% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addin.R
\name{rs_fnmate_below}
\alias{rs_fnmate_below}
\title{rs_fnmate_below}
\usage{
rs_fnmate_below(context = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{context}{from the rstudioapi}
}
\value{
nothing.
}
\description{
Generate a function definition below from a function call in RStudio
}
