% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addin.R
\name{rs_fnmate_clip}
\alias{rs_fnmate_clip}
\title{rs_fnmate_clip}
\usage{
rs_fnmate_clip(context = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{context}{from the rstudioapi}
}
\value{
nothing.
}
\description{
Generate function definition and paste it to the clipboard
}
\author{
Miles McBain
}
