% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugonce_at_cursor.R
\name{rstudio_addin_debugonce}
\alias{rstudio_addin_debugonce}
\title{RStudio addin to call [debugonce()] on the symbol at the cursor.}
\usage{
rstudio_addin_debugonce(context = NULL)
}
\arguments{
\item{context}{RStudio API context from
`rstudioapi::getActiveDocumentContext()`.}
}
\description{
For internal use only. Not a user-side function.
}
\keyword{internal}
